refactor(#1100): split internal/config/config.go into per-concern files
Relocates all types and validators from the 2541-line config.go into 10 new
files per the architect's design: tls.go, auth.go, security_headers.go,
ratelimit.go, upstream.go, database.go, telemetry.go, webhooks.go, secrets.go,
sidecar.go. validateEgressConfig relocated into the existing egress.go peer file.
Zero behavioral change — all 99 config tests pass unchanged.
