+++ internal/config/config.go
@@ -3,20 +3,11 @@ package config
 
 import (
 	"fmt"
-	"net"
-	"net/url"
 	"path/filepath"
-	"regexp"
 	"strings"
-	"time"
 
 	"github.com/spf13/viper"
-
-	"github.com/vibewarden/vibewarden/internal/domain/identity"
 )
-
-// hexColorRE matches valid CSS hex color values: #RGB or #RRGGBB.
-var hexColorRE = regexp.MustCompile(`^#([0-9A-Fa-f]{3}|[0-9A-Fa-f]{6})$`)
 
 // Config holds all configuration for VibeWarden.
 // Fields are loaded from vibewarden.yaml and can be overridden by environment variables.
@@ -146,17 +137,6 @@ type Config struct {
 	ProjectRoot string `mapstructure:"-"`
 }
 
-// WatchConfig holds settings for the config file watcher.
-type WatchConfig struct {
-	// Enabled toggles automatic config reload on file changes. Default: true.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Debounce is the duration to wait after the last file change before
-	// triggering a reload, expressed as a Go duration string (e.g. "500ms").
-	// Default: "500ms".
-	Debounce string `mapstructure:"debounce"`
-}
-
 // InternalNetworkName returns the Docker network name used for internal
 // service-to-service communication. When network isolation is enabled it
 // returns "vibewarden-internal"; otherwise it returns the legacy single
@@ -273,1446 +253,6 @@ func (c *Config) ResolvedBuildContext() string {
 	return c.ProjectRoot + "/" + sub
 }
 
-// DatabasePoolConfig holds connection pool settings for PostgreSQL.
-type DatabasePoolConfig struct {
-	// MaxConns is the maximum number of open connections in the pool.
-	// Default: 10.
-	MaxConns int `mapstructure:"max_conns"`
-
-	// MinConns is the minimum number of idle connections kept open.
-	// Default: 2.
-	MinConns int `mapstructure:"min_conns"`
-}
-
-// DatabaseConfig holds PostgreSQL connection settings used for audit logging
-// and other persistence features.
-type DatabaseConfig struct {
-	// URL is a libpq-compatible connection string or URL.
-	// Example: "postgres://user:pass@localhost:5432/vibewarden?sslmode=disable"
-	// Can be set via VIBEWARDEN_DATABASE_URL env var.
-	URL string `mapstructure:"url"`
-
-	// ExternalURL is the connection URL for an external PostgreSQL instance.
-	// When set, the generated Docker Compose omits the local kratos-db container
-	// and uses this URL as the Kratos DSN instead of the local Postgres service.
-	// Must be a valid postgres:// URL.
-	// Example: "postgres://user:pass@db.example.com:5432/kratos?sslmode=require"
-	// Can be set via VIBEWARDEN_DATABASE_EXTERNAL_URL env var.
-	ExternalURL string `mapstructure:"external_url"`
-
-	// TLSMode controls the PostgreSQL SSL/TLS negotiation mode when building
-	// a DSN. Accepted values: "disable", "require", "verify-ca", "verify-full".
-	// Default: "require".
-	// This value is appended as sslmode=<value> when building the DSN unless
-	// the URL already contains an sslmode query parameter.
-	TLSMode string `mapstructure:"tls_mode"`
-
-	// Pool holds connection pool settings.
-	Pool DatabasePoolConfig `mapstructure:"pool"`
-
-	// ConnectTimeout is the maximum time to wait when establishing a new
-	// Postgres connection, expressed as a Go duration string (e.g. "10s", "30s").
-	// Default: "10s".
-	ConnectTimeout string `mapstructure:"connect_timeout"`
-}
-
-// BuildDSN returns the external URL with connection resilience parameters
-// (sslmode, connect_timeout, pool_max_conns) appended as query parameters.
-// Parameters already present in ExternalURL are not overwritten.
-// Returns an empty string when ExternalURL is empty.
-func (d DatabaseConfig) BuildDSN() string {
-	if d.ExternalURL == "" {
-		return ""
-	}
-
-	u, err := url.Parse(d.ExternalURL)
-	if err != nil {
-		// ExternalURL has already been validated; this path should not be reached.
-		return d.ExternalURL
-	}
-
-	q := u.Query()
-
-	if _, ok := q["sslmode"]; !ok {
-		mode := d.TLSMode
-		if mode == "" {
-			mode = "require"
-		}
-		q.Set("sslmode", mode)
-	}
-
-	if _, ok := q["connect_timeout"]; !ok {
-		timeout := d.ConnectTimeout
-		if timeout == "" {
-			timeout = "10s"
-		}
-		// Postgres connect_timeout is in seconds (integer). Strip the "s" suffix if present.
-		secs := strings.TrimSuffix(timeout, "s")
-		q.Set("connect_timeout", secs)
-	}
-
-	if _, ok := q["pool_max_conns"]; !ok {
-		maxConns := d.Pool.MaxConns
-		if maxConns <= 0 {
-			maxConns = 10
-		}
-		q.Set("pool_max_conns", fmt.Sprintf("%d", maxConns))
-	}
-
-	u.RawQuery = q.Encode()
-	return u.String()
-}
-
-// ResolveURL returns the database connection URL to use for migrations and
-// other direct database access. It prefers the explicit URL field; when that
-// is empty it falls back to BuildDSN (which derives a URL from ExternalURL
-// with resilience parameters appended). Returns an empty string when neither
-// URL nor ExternalURL is configured.
-func (d DatabaseConfig) ResolveURL() string {
-	if d.URL != "" {
-		return d.URL
-	}
-	return d.BuildDSN()
-}
-
-// ServerConfig holds server-related settings.
-type ServerConfig struct {
-	// Host to bind to (default: "127.0.0.1")
-	Host string `mapstructure:"host"`
-	// Port to listen on (default: 8443)
-	Port int `mapstructure:"port"`
-
-	// ReadTimeout is the maximum duration for reading the entire incoming
-	// request including body, expressed as a Go duration string (e.g. "30s").
-	// A value of "0" or "" disables the timeout (no limit).
-	// Default: "30s".
-	ReadTimeout string `mapstructure:"read_timeout"`
-
-	// WriteTimeout is the maximum duration before timing out writes of the
-	// response, expressed as a Go duration string (e.g. "60s").
-	// A value of "0" or "" disables the timeout (no limit).
-	// Default: "60s".
-	WriteTimeout string `mapstructure:"write_timeout"`
-
-	// IdleTimeout is the maximum amount of time to wait for the next request
-	// when keep-alives are enabled, expressed as a Go duration string (e.g. "120s").
-	// A value of "0" or "" disables the timeout (no limit).
-	// Default: "120s".
-	IdleTimeout string `mapstructure:"idle_timeout"`
-}
-
-// UpstreamConfig holds settings for the upstream application being protected.
-type UpstreamConfig struct {
-	// Host of the upstream application (default: "127.0.0.1")
-	Host string `mapstructure:"host"`
-	// Port of the upstream application (default: 3000)
-	Port int `mapstructure:"port"`
-	// Health configures the active upstream health checker.
-	Health UpstreamHealthConfig `mapstructure:"health"`
-}
-
-// UpstreamHealthConfig holds settings for the active upstream health checker.
-type UpstreamHealthConfig struct {
-	// Enabled toggles the background health checker (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Path is the HTTP path probed on the upstream (default: "/health").
-	// Must return a 2xx status code within Timeout to be considered healthy.
-	Path string `mapstructure:"path"`
-
-	// Interval is the time between consecutive probes, as a duration string
-	// (e.g. "10s", "1m"). Default: "10s".
-	Interval string `mapstructure:"interval"`
-
-	// Timeout is the maximum time to wait for a probe response, as a duration
-	// string (e.g. "5s"). Default: "5s".
-	Timeout string `mapstructure:"timeout"`
-
-	// UnhealthyThreshold is the number of consecutive failures required to
-	// transition to Unhealthy. Default: 3.
-	UnhealthyThreshold int `mapstructure:"unhealthy_threshold"`
-
-	// HealthyThreshold is the number of consecutive successes required to
-	// transition to Healthy. Default: 2.
-	HealthyThreshold int `mapstructure:"healthy_threshold"`
-}
-
-// AppConfig configures the user's application in the generated Docker Compose.
-// The default is to use a pre-built image (Image field). Build-from-source is
-// opt-in via the Build field. When both are set, Build takes precedence.
-// When neither is set, no app service is rendered and VibeWarden falls back
-// to forwarding to host.docker.internal.
-type AppConfig struct {
-	// Build is the Docker build context path (e.g., "." for the current directory).
-	// Opt-in: set this only when you want Compose to build the image from source.
-	// When set, the app service is rendered with a build: context directive.
-	Build string `mapstructure:"build"`
-
-	// Image is the Docker image reference (e.g., "myapp:latest" or "ghcr.io/org/myapp:latest").
-	// Default: derived from the project name by `vibew init`/`vibew wrap` (e.g., "myapp:latest").
-	// Can be overridden at runtime via the VIBEWARDEN_APP_IMAGE environment variable.
-	Image string `mapstructure:"image"`
-
-	// Healthcheck is the Docker healthcheck command for the app container.
-	// When empty, the default probe is chosen based on Language:
-	//   python     → python -c "import urllib.request; ..."
-	//   typescript → node -e "require('http').get(...)"
-	//   go/kotlin  → wget -q --spider ...  (Alpine images ship wget)
-	//   (default)  → wget -q --spider ...
-	// Set to "none" to disable the healthcheck entirely.
-	Healthcheck string `mapstructure:"healthcheck"`
-
-	// Language is the app's primary language/runtime. Used to select the
-	// appropriate Docker health check probe when Healthcheck is not set.
-	// Values: "go", "python", "typescript", "kotlin", or empty (auto-detect
-	// at init/wrap time; falls back to wget).
-	Language string `mapstructure:"language"`
-
-	// Environment is a map of custom environment variables injected into the
-	// app container in the generated Docker Compose file. Use this for runtime
-	// configuration (e.g. DATABASE_URL, API keys). Values are rendered as
-	// KEY=VALUE entries in the environment list of the app service.
-	Environment map[string]string `mapstructure:"environment"`
-}
-
-// TLSCertMonitoringConfig holds configuration for the certificate expiry monitor.
-type TLSCertMonitoringConfig struct {
-	// Enabled toggles the certificate expiry monitor (default: true when TLS is enabled).
-	Enabled bool `mapstructure:"enabled"`
-	// CheckInterval is how often the monitor checks certificate expiry (default: 6h).
-	CheckInterval string `mapstructure:"check_interval"`
-	// WarningThreshold is the time-before-expiry that triggers a warning event (default: 720h = 30 days).
-	WarningThreshold string `mapstructure:"warning_threshold"`
-	// CriticalThreshold is the time-before-expiry that triggers a critical event and degraded health (default: 168h = 7 days).
-	CriticalThreshold string `mapstructure:"critical_threshold"`
-}
-
-// TLSConfig holds TLS-related settings.
-type TLSConfig struct {
-	// Enabled toggles TLS (default: false for local dev)
-	Enabled bool `mapstructure:"enabled"`
-	// Domain for TLS certificate (required if enabled with provider "letsencrypt")
-	Domain string `mapstructure:"domain"`
-	// Provider: "letsencrypt" (or alias "acme"), "self-signed", or "external"
-	Provider string `mapstructure:"provider"`
-	// CertPath is the path to a PEM-encoded certificate file.
-	// Required when Provider is "external".
-	CertPath string `mapstructure:"cert_path"`
-	// KeyPath is the path to a PEM-encoded private key file.
-	// Required when Provider is "external".
-	KeyPath string `mapstructure:"key_path"`
-	// StoragePath is the directory where Caddy stores ACME certificates.
-	// Only applies when Provider is "letsencrypt".
-	StoragePath string `mapstructure:"storage_path"`
-	// Email is the ACME account email address used for certificate expiry
-	// notifications and automatic EAB registration with CAs that require it
-	// (e.g. ZeroSSL). Optional for Let's Encrypt.
-	Email string `mapstructure:"email"`
-	// ACMECA is the ACME directory URL to use instead of Let's Encrypt production.
-	// Only applies when Provider is "letsencrypt".
-	// Example: "https://acme-staging-v02.api.letsencrypt.org/directory"
-	ACMECA string `mapstructure:"acme_ca"`
-	// CertMonitoring holds configuration for the background certificate expiry monitor.
-	CertMonitoring TLSCertMonitoringConfig `mapstructure:"cert_monitoring"`
-	// SkipRateLimitCheck disables the Let's Encrypt rate-limit preflight check
-	// run by "vibew doctor" when provider is "letsencrypt". Set to true when
-	// you are intentionally re-issuing within a 168-hour window (e.g. after
-	// certificate revocation) or when crt.sh is unavailable and you want to
-	// suppress the WARN. The --skip-le-preflight flag on "vibew doctor" is the
-	// single-invocation equivalent. Both flags are frozen by ADR-090.
-	SkipRateLimitCheck bool `mapstructure:"skip_rate_limit_check"`
-}
-
-// KratosSMTPConfig holds SMTP settings used by Ory Kratos to send emails.
-type KratosSMTPConfig struct {
-	// Host is the SMTP server hostname (default: "localhost").
-	Host string `mapstructure:"host"`
-	// Port is the SMTP server port (default: 1025).
-	Port int `mapstructure:"port"`
-	// From is the sender address for Kratos emails (default: "no-reply@vibewarden.local").
-	From string `mapstructure:"from"`
-}
-
-// KratosConfig holds Ory Kratos connection settings.
-// These values are used both for the auth middleware and for generating
-// the Kratos config file under .vibewarden/generated/.
-type KratosConfig struct {
-	// PublicURL is the Kratos public API URL (default: "http://127.0.0.1:4433")
-	PublicURL string `mapstructure:"public_url"`
-	// AdminURL is the Kratos admin API URL (default: "http://127.0.0.1:4434")
-	AdminURL string `mapstructure:"admin_url"`
-	// DSN is the data source name for the Kratos database.
-	// Example: "postgres://kratos:secret@localhost:5432/kratos?sslmode=disable"
-	DSN string `mapstructure:"dsn"`
-	// SMTP holds email delivery settings for Kratos.
-	SMTP KratosSMTPConfig `mapstructure:"smtp"`
-	// External indicates that VibeWarden should connect to a user-managed Kratos
-	// instance instead of spinning up a local one.
-	// When true, the generated Docker Compose file omits the kratos,
-	// kratos-migrate, kratos-db, and seed-users containers.
-	// Requires PublicURL and AdminURL to be set to the external instance URLs.
-	// Default: false.
-	External bool `mapstructure:"external"`
-}
-
-// SupportedSocialProviders is the set of accepted provider names for social login.
-// The special value "oidc" indicates a generic OpenID Connect provider.
-var SupportedSocialProviders = map[string]bool{
-	"google":    true,
-	"github":    true,
-	"apple":     true,
-	"facebook":  true,
-	"microsoft": true,
-	"gitlab":    true,
-	"discord":   true,
-	"slack":     true,
-	"spotify":   true,
-	"oidc":      true,
-}
-
-// SocialProviderConfig holds OAuth2/OIDC settings for a single social login provider.
-// It is used as an element of AuthConfig.SocialProviders.
-type SocialProviderConfig struct {
-	// Provider is the provider name.
-	// Accepted values: google, github, apple, facebook, microsoft, gitlab, discord, slack, spotify, oidc.
-	Provider string `mapstructure:"provider"`
-
-	// ClientID is the OAuth2 client ID issued by the provider. Required.
-	ClientID string `mapstructure:"client_id"`
-
-	// ClientSecret is the OAuth2 client secret issued by the provider. Required.
-	// Supports environment variable substitution via ${VAR} syntax in the YAML file.
-	ClientSecret string `mapstructure:"client_secret"`
-
-	// Scopes is an optional list of OAuth2 scopes to request.
-	// When empty, provider-specific defaults are used.
-	Scopes []string `mapstructure:"scopes"`
-
-	// Label is an optional custom label shown on the login button (e.g. "Sign in with Acme").
-	Label string `mapstructure:"label"`
-
-	// TeamID is the Apple Developer Team ID. Required when Provider is "apple".
-	TeamID string `mapstructure:"team_id"`
-
-	// KeyID is the Apple private key ID. Required when Provider is "apple".
-	KeyID string `mapstructure:"key_id"`
-
-	// ID is the unique identifier for the OIDC provider entry (e.g. "acme-oidc").
-	// Required when Provider is "oidc".
-	ID string `mapstructure:"id"`
-
-	// IssuerURL is the OIDC issuer URL (e.g. "https://accounts.google.com").
-	// Required when Provider is "oidc".
-	IssuerURL string `mapstructure:"issuer_url"`
-}
-
-// AuthUIConfig holds theme and URL settings for the built-in authentication UI.
-// It configures the visual appearance of the login, registration, recovery, and
-// settings pages rendered by VibeWarden, as well as the optional custom URL
-// overrides used when mode is "custom".
-type AuthUIConfig struct {
-	// Mode selects whether VibeWarden renders its own auth pages or defers to
-	// custom URLs. Accepted values: "built-in" (default) or "custom".
-	Mode string `mapstructure:"mode"`
-
-	// AppName is the application name shown on the built-in login page.
-	AppName string `mapstructure:"app_name"`
-
-	// LogoURL is an optional URL to a logo image displayed on built-in pages.
-	LogoURL string `mapstructure:"logo_url"`
-
-	// PrimaryColor is the accent color used on built-in pages (hex, default: "#7C3AED").
-	PrimaryColor string `mapstructure:"primary_color"`
-
-	// BackgroundColor is the page background color for built-in pages (hex, default: "#1a1a2e").
-	BackgroundColor string `mapstructure:"background_color"`
-
-	// LoginURL is the URL of the custom login page.
-	// Required when Mode is "custom".
-	LoginURL string `mapstructure:"login_url"`
-
-	// RegistrationURL is the URL of the custom registration page.
-	// Only used when Mode is "custom".
-	RegistrationURL string `mapstructure:"registration_url"`
-
-	// SettingsURL is the URL of the custom account settings page.
-	// Only used when Mode is "custom".
-	SettingsURL string `mapstructure:"settings_url"`
-
-	// RecoveryURL is the URL of the custom account recovery page.
-	// Only used when Mode is "custom".
-	RecoveryURL string `mapstructure:"recovery_url"`
-}
-
-// AuthMode selects the active authentication strategy for incoming requests.
-type AuthMode string
-
-const (
-	// AuthModeKratos uses Ory Kratos session-cookie authentication.
-	// Users must opt in explicitly by setting auth.mode: "kratos".
-	AuthModeKratos AuthMode = "kratos"
-
-	// AuthModeJWT uses JWT/OIDC Bearer token authentication.
-	// Configure the jwt.* sub-section when using this mode.
-	AuthModeJWT AuthMode = "jwt"
-
-	// AuthModeAPIKey uses API key header authentication.
-	AuthModeAPIKey AuthMode = "api-key"
-
-	// AuthModeNone disables authentication entirely. Use only in trusted
-	// environments or when authentication is handled upstream.
-	AuthModeNone AuthMode = "none"
-)
-
-// JWTConfig holds JWT/OIDC authentication settings.
-// It is used when auth.mode is "jwt".
-type JWTConfig struct {
-	// JWKSURL is the URL to fetch the JSON Web Key Set.
-	// Mutually exclusive with IssuerURL: if both are set, JWKSURL takes precedence.
-	// Example: "https://example.auth0.com/.well-known/jwks.json"
-	JWKSURL string `mapstructure:"jwks_url"`
-
-	// IssuerURL is the OIDC issuer URL for auto-discovery.
-	// When set (and JWKSURL is empty), the JWKS URL is discovered from
-	// /.well-known/openid-configuration.
-	// Example: "https://example.auth0.com/"
-	IssuerURL string `mapstructure:"issuer_url"`
-
-	// Issuer is the expected "iss" claim value.
-	// Required when mode is "jwt".
-	// Example: "https://example.auth0.com/"
-	Issuer string `mapstructure:"issuer"`
-
-	// Audience is the expected "aud" claim value.
-	// Required when mode is "jwt".
-	// Example: "my-api"
-	Audience string `mapstructure:"audience"`
-
-	// ClaimsToHeaders maps JWT claim names to HTTP header names.
-	// The mapped claims are injected into requests forwarded to the upstream app.
-	// Default: {"sub": "X-User-Id", "email": "X-User-Email", "email_verified": "X-User-Verified"}
-	// Example:
-	//   claims_to_headers:
-	//     name: X-User-Name
-	//     roles: X-User-Roles
-	ClaimsToHeaders map[string]string `mapstructure:"claims_to_headers"`
-
-	// AllowedAlgorithms restricts which signing algorithms are accepted.
-	// Default: ["RS256", "ES256"].
-	// Never include "none" or symmetric algorithms (HS256) in production.
-	AllowedAlgorithms []string `mapstructure:"allowed_algorithms"`
-
-	// CacheTTL is how long to cache the JWKS before refreshing.
-	// Accepts Go duration strings (e.g. "1h", "30m"). Default: "1h".
-	CacheTTL time.Duration `mapstructure:"cache_ttl"`
-}
-
-// AuthAPIKeyConfig holds settings specific to the API key authentication mode.
-type AuthAPIKeyConfig struct {
-	// Header is the request header from which the API key is extracted.
-	// Defaults to "X-API-Key" when empty.
-	Header string `mapstructure:"header"`
-
-	// Keys is the list of pre-shared API keys recognized by the validator.
-	// Each entry must supply a name, a SHA-256 hex hash of the plaintext key,
-	// and an optional list of scopes.
-	Keys []APIKeyEntry `mapstructure:"keys"`
-
-	// OpenBaoPath is the KV path inside OpenBao where API keys are stored.
-	// When set, the OpenBao adapter is used instead of the static config adapter.
-	// The KV secret at this path must contain string fields whose keys are the
-	// key names and whose values are the SHA-256 hex hashes of the plaintext keys.
-	// Example: openbao_path: "auth/api-keys"
-	OpenBaoPath string `mapstructure:"openbao_path"`
-
-	// CacheTTL is how long the keys fetched from OpenBao are held in memory
-	// before the next refresh. Accepts Go duration strings (e.g. "5m", "1h").
-	// Defaults to "5m". Ignored when OpenBaoPath is empty.
-	CacheTTL string `mapstructure:"cache_ttl"`
-
-	// ScopeRules is an ordered list of path+method authorization rules applied
-	// after successful key validation. The first matching rule determines the
-	// required scopes. When no rule matches, the request is allowed.
-	//
-	// Example:
-	//   scope_rules:
-	//     - path: "/api/v1/*"
-	//       methods: [GET, HEAD]
-	//       required_scopes: ["read"]
-	//     - path: "/admin/*"
-	//       required_scopes: ["admin"]
-	ScopeRules []ScopeRuleConfig `mapstructure:"scope_rules"`
-}
-
-// ScopeRuleConfig describes a single scope-based authorization rule for API
-// key requests. Rules are evaluated in order; the first match wins.
-type ScopeRuleConfig struct {
-	// Path is a glob pattern (stdlib path.Match syntax) matched against the
-	// request URL path. Example: "/api/v1/*"
-	Path string `mapstructure:"path"`
-
-	// Methods is the set of HTTP methods this rule applies to (e.g. ["GET",
-	// "HEAD"]). When empty, the rule applies to all HTTP methods.
-	Methods []string `mapstructure:"methods"`
-
-	// RequiredScopes is the set of scope strings the API key must possess for
-	// the request to be permitted.
-	RequiredScopes []string `mapstructure:"required_scopes"`
-}
-
-// APIKeyEntry represents a single registered API key in the configuration.
-type APIKeyEntry struct {
-	// Name is a human-readable label for the key (e.g. "ci-deploy").
-	Name string `mapstructure:"name"`
-
-	// Hash is the hex-encoded SHA-256 digest of the plaintext key.
-	// Generate with: echo -n "<key>" | sha256sum
-	Hash string `mapstructure:"hash"`
-
-	// Scopes is an optional list of permission scopes granted to this key.
-	Scopes []string `mapstructure:"scopes"`
-}
-
-// AuthConfig holds auth middleware settings.
-//
-// Mode is the single source of truth for whether authentication is on.
-// Set Mode to "none" (or leave it empty) to disable authentication entirely;
-// set it to "kratos", "jwt", or "api-key" to enable the matching strategy.
-// Use the Active method to derive the on/off state at call sites.
-//
-// The legacy auth.enabled field was removed in v0.11.0 per ADR-065. The
-// config loader explicitly rejects any YAML that still sets it.
-type AuthConfig struct {
-	// Mode selects the authentication strategy.
-	// Accepted values: "none" (default), "kratos", "jwt", "api-key".
-	// "none" disables all authentication — use only in trusted environments.
-	// "kratos" activates Ory Kratos session-cookie authentication.
-	// When empty, "none" is used.
-	Mode AuthMode `mapstructure:"mode"`
-
-	// JWT holds settings used when Mode is "jwt".
-	JWT JWTConfig `mapstructure:"jwt"`
-
-	// APIKey holds settings used when Mode is "api-key".
-	APIKey AuthAPIKeyConfig `mapstructure:"api_key"`
-
-	// IdentitySchema selects the identity schema to use.
-	// Accepted values: "email_password" (default), "email_only", "username_password",
-	// "social", or a filesystem path to a custom JSON schema file.
-	// When social_providers are configured and this field is left at its default
-	// ("email_password"), the generate service automatically upgrades to the
-	// "social" schema so that name and picture traits are available.
-	IdentitySchema string `mapstructure:"identity_schema"`
-
-	// PublicPaths is a list of URL path glob patterns that bypass auth.
-	// The /_vibewarden/* prefix is always public (added automatically).
-	// Supports * for single-segment wildcards (e.g. "/static/*").
-	PublicPaths []string `mapstructure:"public_paths"`
-
-	// SessionCookieName is the name of the Kratos session cookie.
-	// Defaults to "ory_kratos_session".
-	SessionCookieName string `mapstructure:"session_cookie_name"`
-
-	// LoginURL is the redirect destination for unauthenticated users.
-	// Defaults to "/self-service/login/browser" when empty.
-	LoginURL string `mapstructure:"login_url"`
-
-	// OnKratosUnavailable controls behavior when Kratos cannot be reached.
-	// Accepted values:
-	//   "503"          (default) — return 503 for all protected requests (fail-closed).
-	//   "allow_public" — serve requests to public paths; block protected paths with 503.
-	OnKratosUnavailable string `mapstructure:"on_kratos_unavailable"`
-
-	// SocialProviders is a list of OAuth2/OIDC social login providers to enable.
-	// Each entry requires at minimum a provider name, client_id, and client_secret.
-	SocialProviders []SocialProviderConfig `mapstructure:"social_providers"`
-
-	// RolePaths maps role names to URL path patterns that require that role.
-	// When configured, authenticated users whose Kratos identity trait "role"
-	// does not match the required role for a path receive HTTP 403 Forbidden.
-	// Only valid when Mode is "kratos".
-	//
-	// Example:
-	//   role_paths:
-	//     admin:
-	//       - /admin/*
-	//     moderator:
-	//       - /admin/moderation/*
-	RolePaths map[string][]string `mapstructure:"role_paths"`
-
-	// UI holds theme and URL settings for the built-in or custom auth pages.
-	UI AuthUIConfig `mapstructure:"ui"`
-}
-
-// Active reports whether the authentication middleware should be active.
-//
-// Auth is active when Mode is a non-empty, non-"none" value. An empty Mode
-// (which viper normalises to AuthModeNone via SetDefault) and AuthModeNone
-// both mean "auth disabled".
-//
-// This helper is the single derivation for "is auth on" and every consumer
-// that needs that signal should call it rather than testing Mode directly.
-// See ADR-065.
-func (a AuthConfig) Active() bool {
-	return a.Mode != "" && a.Mode != AuthModeNone
-}
-
-// RateLimitConfig holds rate limiting settings.
-type RateLimitConfig struct {
-	// Enabled toggles rate limiting (default: true)
-	Enabled bool `mapstructure:"enabled"`
-
-	// Store selects the backing store for limiter state.
-	// Accepted values: "memory" (default), "redis".
-	// "memory" uses a per-process token bucket — no external dependencies.
-	// "redis" uses a Redis-backed distributed token bucket.
-	Store string `mapstructure:"store"`
-
-	// Redis holds connection settings for the Redis store.
-	// Only used when Store is "redis".
-	Redis RateLimitRedisConfig `mapstructure:"redis"`
-
-	// PerIP configures per-IP rate limits applied to all requests.
-	PerIP RateLimitRuleConfig `mapstructure:"per_ip"`
-
-	// PerUser configures per-user rate limits applied to authenticated requests only.
-	PerUser RateLimitRuleConfig `mapstructure:"per_user"`
-
-	// TrustProxyHeaders enables reading X-Forwarded-For to determine the real client IP.
-	// Only enable when VibeWarden is behind a trusted reverse proxy.
-	TrustProxyHeaders bool `mapstructure:"trust_proxy_headers"`
-
-	// ExemptPaths is a list of glob patterns for paths that bypass rate limiting.
-	// The /_vibewarden/* prefix is always exempt and is added automatically.
-	ExemptPaths []string `mapstructure:"exempt_paths"`
-}
-
-// RateLimitRedisConfig holds Redis connection settings for the rate limit store.
-type RateLimitRedisConfig struct {
-	// URL is the Redis connection URL (e.g. "redis://:password@localhost:6379/0"
-	// or "rediss://user:password@redis.example.com:6380/1" for TLS).
-	// When set, URL takes precedence over Address, Password, and DB.
-	// Use a redis:// URL for unencrypted connections and a rediss:// URL for TLS.
-	// Optional: if omitted, Address is used to build the connection.
-	URL string `mapstructure:"url"`
-
-	// Address is the Redis server address in host:port form (default: "localhost:6379").
-	// Used when URL is empty. At least one of URL or Address is required when
-	// rate_limit.store is "redis".
-	Address string `mapstructure:"address"`
-
-	// Password is the Redis AUTH password (default: empty, no auth).
-	// Ignored when URL is set (embed credentials in the URL instead).
-	Password string `mapstructure:"password"`
-
-	// DB is the Redis logical database index (default: 0).
-	// Ignored when URL is set (embed the DB index in the URL path instead).
-	DB int `mapstructure:"db"`
-
-	// PoolSize is the maximum number of socket connections held in the pool
-	// (default: 0, which lets go-redis choose a sensible value based on CPU count).
-	PoolSize int `mapstructure:"pool_size"`
-
-	// KeyPrefix is the namespace prefix prepended to every Redis key
-	// (default: "vibewarden").
-	KeyPrefix string `mapstructure:"key_prefix"`
-
-	// Fallback controls whether the rate limiter falls back to the in-memory
-	// store when Redis is unavailable (default: true — fail-open).
-	// Set to false to enable fail-closed mode: requests are denied when
-	// Redis is unreachable.
-	Fallback bool `mapstructure:"fallback"`
-
-	// HealthCheckInterval is how often the background goroutine probes Redis
-	// for recovery after a failure, expressed as a duration string (e.g. "30s").
-	// Default: "30s".
-	HealthCheckInterval string `mapstructure:"health_check_interval"`
-}
-
-// HasExternalURL reports whether an explicit Redis URL has been configured.
-// When true, VibeWarden connects to that external Redis instance directly and
-// the generated Docker Compose file omits the local redis container.
-func (r RateLimitRedisConfig) HasExternalURL() bool {
-	return r.URL != ""
-}
-
-// RateLimitRuleConfig holds the sustained rate and burst size for a rate limit.
-type RateLimitRuleConfig struct {
-	// RequestsPerSecond is the sustained request rate (default: 10 for IP, 100 for user).
-	RequestsPerSecond float64 `mapstructure:"requests_per_second"`
-
-	// Burst is the maximum number of requests allowed in a burst
-	// above the sustained rate (default: 20 for IP, 200 for user).
-	Burst int `mapstructure:"burst"`
-}
-
-// LogConfig holds logging settings.
-type LogConfig struct {
-	// Level: "debug", "info", "warn", "error" (default: "info")
-	Level string `mapstructure:"level"`
-	// Format: "json" or "text" (default: "json")
-	Format string `mapstructure:"format"`
-	// AccessLog enables access logging of every completed HTTP request (default: true).
-	// When true, each request is logged at INFO level with method, path, status,
-	// duration, client IP, request ID, user agent, and bytes written.
-	AccessLog bool `mapstructure:"access_log"`
-}
-
-// AdminConfig holds admin API settings.
-type AdminConfig struct {
-	// Enabled toggles the admin API (default: false)
-	Enabled bool `mapstructure:"enabled"`
-	// Token is the bearer token for admin API authentication.
-	// Can be set via VIBEWARDEN_ADMIN_TOKEN env var.
-	Token string `mapstructure:"token"`
-}
-
-// SecurityHeadersConfig holds security header settings.
-type SecurityHeadersConfig struct {
-	// Enabled toggles security headers middleware (default: true)
-	Enabled bool `mapstructure:"enabled"`
-
-	// HSTSMaxAge is the Strict-Transport-Security max-age in seconds (default: 31536000 = 1 year)
-	HSTSMaxAge int `mapstructure:"hsts_max_age"`
-	// HSTSIncludeSubDomains includes the includeSubDomains directive (default: true)
-	HSTSIncludeSubDomains bool `mapstructure:"hsts_include_subdomains"`
-	// HSTSPreload includes the preload directive (default: false — requires manual submission)
-	HSTSPreload bool `mapstructure:"hsts_preload"`
-
-	// ContentTypeNosniff sets X-Content-Type-Options: nosniff (default: true)
-	ContentTypeNosniff bool `mapstructure:"content_type_nosniff"`
-
-	// FrameOption sets X-Frame-Options value: "DENY", "SAMEORIGIN", or "" to disable (default: "DENY")
-	FrameOption string `mapstructure:"frame_option"`
-
-	// ContentSecurityPolicy sets Content-Security-Policy value.
-	// An empty string (the default) disables the header entirely; users opt in
-	// by setting an explicit policy in vibewarden.yaml.
-	// When both ContentSecurityPolicy and CSP are set, ContentSecurityPolicy
-	// takes precedence for backward compatibility.
-	ContentSecurityPolicy string `mapstructure:"content_security_policy"`
-
-	// CSP holds a structured, declarative Content-Security-Policy configuration.
-	// It is used to generate a Content-Security-Policy header value when
-	// ContentSecurityPolicy is not set. If ContentSecurityPolicy is non-empty it
-	// always takes precedence.
-	CSP CSPConfig `mapstructure:"csp"`
-
-	// ReferrerPolicy sets Referrer-Policy value (default: "strict-origin-when-cross-origin")
-	ReferrerPolicy string `mapstructure:"referrer_policy"`
-
-	// PermissionsPolicy sets Permissions-Policy value (default: "")
-	PermissionsPolicy string `mapstructure:"permissions_policy"`
-
-	// CrossOriginOpenerPolicy sets Cross-Origin-Opener-Policy value (default: "same-origin")
-	CrossOriginOpenerPolicy string `mapstructure:"cross_origin_opener_policy"`
-
-	// CrossOriginResourcePolicy sets Cross-Origin-Resource-Policy value (default: "same-origin")
-	CrossOriginResourcePolicy string `mapstructure:"cross_origin_resource_policy"`
-
-	// PermittedCrossDomainPolicies sets X-Permitted-Cross-Domain-Policies value (default: "none")
-	PermittedCrossDomainPolicies string `mapstructure:"permitted_cross_domain_policies"`
-
-	// SuppressViaHeader removes the Via header from proxied responses (default: true)
-	SuppressViaHeader bool `mapstructure:"suppress_via_header"`
-}
-
-// CSPConfig holds a declarative Content-Security-Policy configuration.
-// Each field maps to a CSP fetch directive and contains the list of allowed
-// sources. Sources are written verbatim, so keyword tokens must include their
-// surrounding single-quotes (e.g. "'self'", "'none'", "'unsafe-inline'").
-// An empty slice means the directive is omitted from the generated header.
-type CSPConfig struct {
-	// DefaultSrc sets the default-src directive.
-	DefaultSrc []string `mapstructure:"default_src"`
-
-	// ScriptSrc sets the script-src directive.
-	ScriptSrc []string `mapstructure:"script_src"`
-
-	// StyleSrc sets the style-src directive.
-	StyleSrc []string `mapstructure:"style_src"`
-
-	// ImgSrc sets the img-src directive.
-	ImgSrc []string `mapstructure:"img_src"`
-
-	// ConnectSrc sets the connect-src directive.
-	ConnectSrc []string `mapstructure:"connect_src"`
-
-	// FontSrc sets the font-src directive.
-	FontSrc []string `mapstructure:"font_src"`
-
-	// FrameSrc sets the frame-src directive.
-	FrameSrc []string `mapstructure:"frame_src"`
-
-	// MediaSrc sets the media-src directive.
-	MediaSrc []string `mapstructure:"media_src"`
-
-	// ObjectSrc sets the object-src directive.
-	ObjectSrc []string `mapstructure:"object_src"`
-
-	// ManifestSrc sets the manifest-src directive.
-	ManifestSrc []string `mapstructure:"manifest_src"`
-
-	// WorkerSrc sets the worker-src directive.
-	WorkerSrc []string `mapstructure:"worker_src"`
-
-	// ChildSrc sets the child-src directive.
-	ChildSrc []string `mapstructure:"child_src"`
-
-	// FormAction sets the form-action directive.
-	FormAction []string `mapstructure:"form_action"`
-
-	// FrameAncestors sets the frame-ancestors directive.
-	FrameAncestors []string `mapstructure:"frame_ancestors"`
-
-	// BaseURI sets the base-uri directive.
-	BaseURI []string `mapstructure:"base_uri"`
-}
-
-// WAFConfig holds Web Application Firewall settings.
-type WAFConfig struct {
-	// Enabled toggles the WAF rule engine (default: true).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Mode controls WAF response to detections: "block" or "detect" (default: "detect").
-	Mode string `mapstructure:"mode"`
-
-	// Rules toggles individual rule categories.
-	Rules WAFRulesConfig `mapstructure:"rules"`
-
-	// ExemptPaths is a list of URL path glob patterns that bypass WAF scanning.
-	ExemptPaths []string `mapstructure:"exempt_paths"`
-
-	// ContentTypeValidation configures the Content-Type validation middleware.
-	ContentTypeValidation ContentTypeValidationConfig `mapstructure:"content_type_validation"`
-}
-
-// WAFRulesConfig toggles individual WAF rule categories.
-type WAFRulesConfig struct {
-	// SQLInjection toggles SQLi detection rules (default: true).
-	SQLInjection bool `mapstructure:"sqli"`
-
-	// XSS toggles cross-site scripting detection rules (default: true).
-	XSS bool `mapstructure:"xss"`
-
-	// PathTraversal toggles path traversal detection rules (default: true).
-	PathTraversal bool `mapstructure:"path_traversal"`
-
-	// CommandInjection toggles command injection detection rules (default: true).
-	CommandInjection bool `mapstructure:"command_injection"`
-}
-
-// ContentTypeValidationConfig holds settings for the Content-Type validation middleware.
-type ContentTypeValidationConfig struct {
-	// Enabled toggles Content-Type validation on body-bearing requests (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Allowed is the list of permitted media types.
-	// Requests with a Content-Type not in this list receive 415 Unsupported Media Type.
-	// Default: ["application/json", "application/x-www-form-urlencoded", "multipart/form-data"]
-	Allowed []string `mapstructure:"allowed"`
-}
-
-// InputValidationConfig holds request input size limit settings.
-type InputValidationConfig struct {
-	// Enabled toggles the input validation middleware (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// MaxURLLength is the maximum allowed length of the raw request URI in bytes
-	// (path + query string). Default: 2048. Zero disables this check.
-	MaxURLLength int `mapstructure:"max_url_length"`
-
-	// MaxQueryStringLength is the maximum allowed query string length in bytes,
-	// not including the leading "?". Default: 2048. Zero disables this check.
-	MaxQueryStringLength int `mapstructure:"max_query_string_length"`
-
-	// MaxHeaderCount is the maximum number of request headers allowed.
-	// Default: 100. Zero disables this check.
-	MaxHeaderCount int `mapstructure:"max_header_count"`
-
-	// MaxHeaderSize is the maximum allowed byte length of any single header
-	// value. Default: 8192. Zero disables this check.
-	MaxHeaderSize int `mapstructure:"max_header_size"`
-
-	// PathOverrides defines per-path limit overrides.
-	// The first entry whose Path glob pattern (path.Match syntax) matches the
-	// request URL path wins. Non-zero fields in the matching entry override the
-	// global limits.
-	PathOverrides []InputValidationPathOverrideConfig `mapstructure:"path_overrides"`
-}
-
-// InputValidationPathOverrideConfig defines per-path limit overrides for the
-// input validation middleware.
-type InputValidationPathOverrideConfig struct {
-	// Path is a glob pattern (path.Match syntax) matched against the request
-	// URL path (e.g. "/api/upload", "/static/*").
-	Path string `mapstructure:"path"`
-
-	// MaxURLLength overrides the global limit for matching paths.
-	// Zero means inherit the global value.
-	MaxURLLength int `mapstructure:"max_url_length"`
-
-	// MaxQueryStringLength overrides the global limit for matching paths.
-	// Zero means inherit the global value.
-	MaxQueryStringLength int `mapstructure:"max_query_string_length"`
-
-	// MaxHeaderCount overrides the global limit for matching paths.
-	// Zero means inherit the global value.
-	MaxHeaderCount int `mapstructure:"max_header_count"`
-
-	// MaxHeaderSize overrides the global limit for matching paths.
-	// Zero means inherit the global value.
-	MaxHeaderSize int `mapstructure:"max_header_size"`
-}
-
-// CORSConfig holds Cross-Origin Resource Sharing settings.
-type CORSConfig struct {
-	// Enabled toggles the CORS plugin (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// AllowedOrigins is the list of origins permitted to make cross-origin
-	// requests. Use ["*"] to allow all origins (development only).
-	AllowedOrigins []string `mapstructure:"allowed_origins"`
-
-	// AllowedMethods is the list of HTTP methods permitted in cross-origin
-	// requests (default: GET, POST, PUT, DELETE, OPTIONS).
-	AllowedMethods []string `mapstructure:"allowed_methods"`
-
-	// AllowedHeaders is the list of request headers permitted in cross-origin
-	// requests (default: Content-Type, Authorization).
-	AllowedHeaders []string `mapstructure:"allowed_headers"`
-
-	// ExposedHeaders is the list of response headers exposed to the browser
-	// via Access-Control-Expose-Headers (default: []).
-	ExposedHeaders []string `mapstructure:"exposed_headers"`
-
-	// AllowCredentials, when true, sets Access-Control-Allow-Credentials: true.
-	// Must not be combined with AllowedOrigins: ["*"] (default: false).
-	AllowCredentials bool `mapstructure:"allow_credentials"`
-
-	// MaxAge is the number of seconds the browser may cache the preflight
-	// response (Access-Control-Max-Age). Zero omits the header (default: 0).
-	MaxAge int `mapstructure:"max_age"`
-}
-
-// MetricsConfig holds Prometheus metrics settings.
-//
-// Deprecated: Use TelemetryConfig instead. MetricsConfig remains only for backward
-// compatibility. Settings are migrated to TelemetryConfig at startup via MigrateLegacyMetrics.
-type MetricsConfig struct {
-	// Enabled toggles metrics collection and the /_vibewarden/metrics endpoint (default: true).
-	Enabled bool `mapstructure:"enabled"`
-
-	// PathPatterns is a list of URL path normalization patterns using :param syntax.
-	// Example: "/users/:id", "/api/v1/items/:item_id/comments/:comment_id"
-	// Paths that don't match any pattern are recorded as "other".
-	PathPatterns []string `mapstructure:"path_patterns"`
-}
-
-// TelemetryConfig holds all telemetry export settings.
-// This replaces the narrower MetricsConfig and supports both pull (Prometheus)
-// and push (OTLP) export modes.
-//
-// Prometheus is the automatic fallback: when no telemetry block is present in
-// vibewarden.yaml, prometheus.enabled defaults to true and otlp.enabled defaults
-// to false. This means /_vibewarden/metrics always works out of the box, with
-// zero configuration required. Existing Prometheus scrapers and Grafana dashboards
-// continue to work without any changes.
-//
-// To add OTLP push export, set otlp.enabled = true and provide an endpoint.
-// Both exporters can run simultaneously.
-type TelemetryConfig struct {
-	// Enabled toggles telemetry collection entirely (default: true).
-	Enabled bool `mapstructure:"enabled"`
-
-	// PathPatterns is a list of URL path normalization patterns using :param syntax.
-	// Example: "/users/:id", "/api/v1/items/:item_id/comments/:comment_id"
-	// Paths that don't match any pattern are recorded as "other".
-	PathPatterns []string `mapstructure:"path_patterns"`
-
-	// Prometheus configures the pull-based Prometheus exporter.
-	Prometheus PrometheusExporterConfig `mapstructure:"prometheus"`
-
-	// OTLP configures the push-based OTLP exporter.
-	OTLP OTLPExporterConfig `mapstructure:"otlp"`
-
-	// Logs configures structured event log export settings.
-	Logs LogsConfig `mapstructure:"logs"`
-
-	// Traces configures distributed tracing settings.
-	Traces TracesConfig `mapstructure:"traces"`
-}
-
-// TracesConfig holds distributed tracing settings.
-type TracesConfig struct {
-	// Enabled toggles distributed tracing (default: false).
-	// When enabled, a span is created for each HTTP request and exported via OTLP.
-	// Requires telemetry.otlp.enabled to be true.
-	Enabled bool `mapstructure:"enabled"`
-}
-
-// LogsConfig holds log export settings.
-type LogsConfig struct {
-	// OTLP toggles OTLP log export (default: false).
-	// When enabled, structured events are exported to the same OTLP endpoint as metrics.
-	// Requires telemetry.otlp.endpoint to be configured.
-	OTLP bool `mapstructure:"otlp"`
-}
-
-// PrometheusExporterConfig configures the Prometheus pull-based exporter.
-type PrometheusExporterConfig struct {
-	// Enabled toggles the Prometheus exporter (default: true).
-	// When enabled, metrics are served at /_vibewarden/metrics.
-	Enabled bool `mapstructure:"enabled"`
-}
-
-// OTLPExporterConfig configures the OTLP push-based exporter.
-type OTLPExporterConfig struct {
-	// Enabled toggles the OTLP exporter (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Endpoint is the OTLP HTTP endpoint URL (e.g., "http://localhost:4318").
-	// Required when Enabled is true.
-	Endpoint string `mapstructure:"endpoint"`
-
-	// Headers are optional HTTP headers for authentication.
-	// Example: {"Authorization": "Bearer <token>"}
-	Headers map[string]string `mapstructure:"headers"`
-
-	// Interval is the export interval as a duration string (default: "30s").
-	// Metrics are batched and pushed at this interval.
-	Interval string `mapstructure:"interval"`
-
-	// Protocol is "http" or "grpc" (default: "http").
-	// Only "http" is supported in this version.
-	Protocol string `mapstructure:"protocol"`
-}
-
-// ResilienceConfig holds resilience settings for upstream connections.
-type ResilienceConfig struct {
-	// Timeout is the maximum time to wait for the upstream application to
-	// respond, expressed as a duration string (e.g. "30s", "1m").
-	// A value of "0" or "" disables the timeout (no limit).
-	// Default: "30s".
-	Timeout string `mapstructure:"timeout"`
-
-	// CircuitBreaker configures the circuit breaker middleware.
-	CircuitBreaker CircuitBreakerConfig `mapstructure:"circuit_breaker"`
-
-	// Retry configures the retry-with-exponential-backoff middleware.
-	Retry RetryConfig `mapstructure:"retry"`
-}
-
-// RetryConfig holds retry-with-exponential-backoff settings.
-type RetryConfig struct {
-	// Enabled toggles the retry middleware (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// MaxAttempts is the total number of attempts including the initial request.
-	// Must be >= 2 when Enabled is true. Default: 3.
-	MaxAttempts int `mapstructure:"max_attempts"`
-
-	// InitialBackoff is the wait before the first retry, as a duration string
-	// (e.g. "100ms", "500ms"). Default: "100ms".
-	InitialBackoff string `mapstructure:"backoff"`
-
-	// MaxBackoff is the upper bound on the computed backoff, as a duration string
-	// (e.g. "10s"). Default: "10s".
-	MaxBackoff string `mapstructure:"max_backoff"`
-
-	// RetryOn is the list of HTTP status codes that trigger a retry.
-	// Default: [502, 503, 504].
-	RetryOn []int `mapstructure:"retry_on"`
-}
-
-// CircuitBreakerConfig holds circuit breaker settings.
-type CircuitBreakerConfig struct {
-	// Enabled toggles the circuit breaker middleware.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Threshold is the number of consecutive failures required to trip the
-	// circuit from Closed to Open. Must be > 0 when Enabled is true.
-	// Default: 5.
-	Threshold int `mapstructure:"threshold"`
-
-	// Timeout is how long the circuit stays Open before transitioning to
-	// HalfOpen to allow a probe request, expressed as a duration string
-	// (e.g. "60s", "1m"). Must be > 0 when Enabled is true.
-	// Default: "60s".
-	Timeout string `mapstructure:"timeout"`
-}
-
-// BodySizeConfig holds request body size limit settings.
-type BodySizeConfig struct {
-	// Max is the global default maximum request body size as a human-readable
-	// string (e.g. "1MB", "512KB"). Parsed at startup.
-	// An empty string or "0" means no limit.
-	Max string `mapstructure:"max"`
-
-	// Overrides defines per-path body size limits.
-	// Each entry can increase or decrease the global limit for a specific path.
-	Overrides []BodySizeOverrideConfig `mapstructure:"overrides"`
-}
-
-// BodySizeOverrideConfig defines a per-path body size limit.
-type BodySizeOverrideConfig struct {
-	// Path is the URL path prefix to match (e.g. "/api/upload").
-	Path string `mapstructure:"path"`
-
-	// Max is the maximum request body size for this path as a human-readable
-	// string (e.g. "50MB"). An empty string or "0" means no limit for this path.
-	Max string `mapstructure:"max"`
-}
-
-// IPFilterConfig holds IP-based access control settings.
-type IPFilterConfig struct {
-	// Enabled toggles the IP filter plugin (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Mode selects the filter behaviour: "allowlist" or "blocklist" (default: "blocklist").
-	// allowlist: only listed IPs/CIDRs may access the service.
-	// blocklist: listed IPs/CIDRs are blocked; all others are permitted.
-	Mode string `mapstructure:"mode"`
-
-	// Addresses is the list of IP addresses or CIDR ranges to match against.
-	// Examples: "10.0.0.0/8", "192.168.1.100", "2001:db8::/32".
-	Addresses []string `mapstructure:"addresses"`
-
-	// TrustProxyHeaders enables reading X-Forwarded-For for the real client IP.
-	// Only enable when VibeWarden runs behind a trusted reverse proxy.
-	TrustProxyHeaders bool `mapstructure:"trust_proxy_headers"`
-}
-
-// WebhooksConfig holds all webhook delivery settings.
-// It maps to the webhooks section of vibewarden.yaml.
-type WebhooksConfig struct {
-	// Endpoints is the list of webhook endpoints to deliver events to.
-	Endpoints []WebhookEndpointConfig `mapstructure:"endpoints"`
-
-	// SignatureVerification configures inbound webhook signature verification.
-	SignatureVerification WebhookSignatureVerificationConfig `mapstructure:"signature_verification"`
-}
-
-// WebhookSignatureVerificationConfig holds all settings for inbound webhook
-// signature verification. It maps to the webhooks.signature_verification
-// section of vibewarden.yaml.
-type WebhookSignatureVerificationConfig struct {
-	// Enabled toggles inbound webhook signature verification (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Paths is the ordered list of per-path signature verification rules.
-	Paths []WebhookSignaturePathConfig `mapstructure:"paths"`
-}
-
-// WebhookSignaturePathConfig holds the per-path webhook signature rule.
-// It maps to a single entry under webhooks.signature_verification.paths.
-type WebhookSignaturePathConfig struct {
-	// Path is the URL path this rule applies to (exact match, required).
-	Path string `mapstructure:"path"`
-
-	// Provider selects the signature format: "stripe", "github", "slack",
-	// "twilio", or "generic". Required.
-	Provider string `mapstructure:"provider"`
-
-	// SecretEnvVar is the name of the environment variable containing the
-	// shared HMAC secret. Required — secrets must not be stored in the config
-	// file directly. Use ${VAR_NAME} or just the variable name.
-	SecretEnvVar string `mapstructure:"secret_env_var"`
-
-	// Header is the custom HTTP header name used when Provider is "generic".
-	// Ignored for all other providers.
-	Header string `mapstructure:"header,omitempty"`
-}
-
-// WebhookEndpointConfig holds the settings for a single webhook endpoint.
-type WebhookEndpointConfig struct {
-	// URL is the HTTP(S) endpoint to POST events to. Required.
-	URL string `mapstructure:"url"`
-
-	// Events is the list of event types to deliver to this endpoint.
-	// Use "*" to subscribe to all events.
-	Events []string `mapstructure:"events"`
-
-	// Format selects the payload format: "raw" (default), "slack", or "discord".
-	Format string `mapstructure:"format"`
-
-	// TimeoutSeconds is the per-request HTTP timeout in seconds (default: 10).
-	TimeoutSeconds int `mapstructure:"timeout_seconds"`
-}
-
-// SecretsConfig holds all settings for the secret management plugin.
-// It maps to the secrets section of vibewarden.yaml.
-type SecretsConfig struct {
-	// Enabled toggles the secrets plugin (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Store selects the secret store backend: "builtin" or "openbao".
-	// Default: "builtin". When empty or unset, "builtin" is used.
-	Store string `mapstructure:"store"`
-
-	// Provider selects the secret store backend (default: "openbao").
-	// Deprecated: use Store instead. Kept for backward compatibility.
-	Provider string `mapstructure:"provider"`
-
-	// Builtin holds settings for the built-in encrypted file store.
-	Builtin SecretsBuiltinConfig `mapstructure:"builtin"`
-
-	// OpenBao holds connection and authentication settings for the OpenBao server.
-	OpenBao SecretsOpenBaoConfig `mapstructure:"openbao"`
-
-	// Inject defines how fetched secrets are surfaced to the upstream app.
-	Inject SecretsInjectConfig `mapstructure:"inject"`
-
-	// Dynamic holds settings for dynamic credential generation.
-	Dynamic SecretsDynamicConfig `mapstructure:"dynamic"`
-
-	// Health holds settings for the secret health check subsystem.
-	Health SecretsHealthConfig `mapstructure:"health"`
-
-	// CacheTTL is how long fetched secrets are held in memory (default: "5m").
-	CacheTTL string `mapstructure:"cache_ttl"`
-}
-
-// SecretsBuiltinConfig holds settings for the built-in encrypted file store.
-type SecretsBuiltinConfig struct {
-	// Path is the location of the encrypted secrets file.
-	// Default: ".vibewarden/secrets.enc".
-	Path string `mapstructure:"path"`
-
-	// KeyFile is the path to a file containing the 32-byte hex-encoded master key.
-	// If empty, the VIBEWARDEN_SECRETS_MASTER_KEY environment variable is used.
-	KeyFile string `mapstructure:"key_file"`
-}
-
-// SecretsOpenBaoConfig holds connection settings for the OpenBao server.
-type SecretsOpenBaoConfig struct {
-	// Address is the OpenBao server URL (e.g. "http://openbao:8200").
-	Address string `mapstructure:"address"`
-
-	// Auth holds the authentication credentials.
-	Auth SecretsOpenBaoAuthConfig `mapstructure:"auth"`
-
-	// MountPath is the KV v2 mount path (default: "secret").
-	MountPath string `mapstructure:"mount_path"`
-}
-
-// SecretsOpenBaoAuthConfig holds authentication credentials for OpenBao.
-type SecretsOpenBaoAuthConfig struct {
-	// Method is the auth method: "token" or "approle".
-	Method string `mapstructure:"method"`
-
-	// Token is the static token. Used when Method is "token".
-	Token string `mapstructure:"token"`
-
-	// RoleID is the AppRole role_id. Used when Method is "approle".
-	RoleID string `mapstructure:"role_id"`
-
-	// SecretID is the AppRole secret_id. Used when Method is "approle".
-	SecretID string `mapstructure:"secret_id"`
-}
-
-// SecretsInjectConfig defines how secrets are injected into proxied requests.
-type SecretsInjectConfig struct {
-	// Headers is the list of secrets to inject as HTTP request headers.
-	Headers []SecretsHeaderInjection `mapstructure:"headers"`
-
-	// EnvFile is the path to write a .env file containing secret values.
-	EnvFile string `mapstructure:"env_file"`
-
-	// Env is the list of secrets to write into the env file.
-	Env []SecretsEnvInjection `mapstructure:"env"`
-}
-
-// SecretsHeaderInjection maps a secret key to an HTTP request header name.
-type SecretsHeaderInjection struct {
-	// SecretPath is the KV path of the secret.
-	SecretPath string `mapstructure:"secret_path"`
-
-	// SecretKey is the key within the secret map.
-	SecretKey string `mapstructure:"secret_key"`
-
-	// Header is the HTTP header name to set.
-	Header string `mapstructure:"header"`
-
-	// ValueTemplate is an optional template string containing ${secret://path/key}
-	// placeholders. When set, the resolved template is used as the header value
-	// instead of the raw secret. When empty, the secret value is used directly.
-	ValueTemplate string `mapstructure:"value_template"`
-}
-
-// SecretsEnvInjection maps a secret key to an environment variable name.
-type SecretsEnvInjection struct {
-	// SecretPath is the KV path of the secret.
-	SecretPath string `mapstructure:"secret_path"`
-
-	// SecretKey is the key within the secret map.
-	SecretKey string `mapstructure:"secret_key"`
-
-	// EnvVar is the environment variable name to write in the env file.
-	EnvVar string `mapstructure:"env_var"`
-
-	// ValueTemplate is an optional template string containing ${secret://path/key}
-	// placeholders. When set, the resolved template is used as the env var value
-	// instead of the raw secret. When empty, the secret value is used directly.
-	ValueTemplate string `mapstructure:"value_template"`
-}
-
-// SecretsDynamicConfig holds settings for dynamic credential generation.
-type SecretsDynamicConfig struct {
-	// Postgres configures the OpenBao database engine for Postgres dynamic creds.
-	Postgres SecretsDynamicPostgresConfig `mapstructure:"postgres"`
-}
-
-// SecretsDynamicPostgresConfig holds settings for dynamic Postgres credential generation.
-type SecretsDynamicPostgresConfig struct {
-	// Enabled toggles dynamic Postgres credential generation.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Roles is the list of OpenBao database roles to request credentials for.
-	Roles []SecretsDynamicRole `mapstructure:"roles"`
-}
-
-// SecretsDynamicRole defines a single OpenBao database role and where to inject its credentials.
-type SecretsDynamicRole struct {
-	// Name is the OpenBao database role name.
-	Name string `mapstructure:"name"`
-
-	// EnvVarUser is the env var name to write the generated username into.
-	EnvVarUser string `mapstructure:"env_var_user"`
-
-	// EnvVarPassword is the env var name to write the generated password into.
-	EnvVarPassword string `mapstructure:"env_var_password"`
-}
-
-// SecretsHealthConfig holds settings for the secret health check subsystem.
-type SecretsHealthConfig struct {
-	// CheckInterval is how often to run health checks (default: "6h").
-	CheckInterval string `mapstructure:"check_interval"`
-
-	// MaxStaticAge is the maximum acceptable age of a static secret (default: "2160h").
-	MaxStaticAge string `mapstructure:"max_static_age"`
-
-	// WeakPatterns is the list of substrings that indicate a weak/default secret.
-	WeakPatterns []string `mapstructure:"weak_patterns"`
-}
-
-// OverridesConfig provides escape hatches for users who need to supply
-// hand-crafted configuration files instead of relying on VibeWarden's
-// auto-generation. All fields are optional.
-type OverridesConfig struct {
-	// KratosConfig is the path to a custom kratos.yml file.
-	// When non-empty, VibeWarden uses this file instead of generating one.
-	KratosConfig string `mapstructure:"kratos_config"`
-
-	// ComposeFile is the path to a custom docker-compose.yml file.
-	// When non-empty, VibeWarden uses this file instead of generating one.
-	ComposeFile string `mapstructure:"compose_file"`
-
-	// IdentitySchema is the path to a custom Kratos identity schema JSON file.
-	// When non-empty, this file is used instead of the preset selected by auth.identity_schema.
-	IdentitySchema string `mapstructure:"identity_schema"`
-}
-
-// ObservabilityConfig holds settings for the optional observability stack.
-// When enabled, vibewarden generate produces Prometheus, Grafana, Loki, and
-// Promtail configs under .vibewarden/generated/observability/.
-type ObservabilityConfig struct {
-	// Enabled toggles generation of the observability stack (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// GrafanaPort is the host port Grafana binds to (default: 3001).
-	// This avoids conflict with common app ports like 3000.
-	GrafanaPort int `mapstructure:"grafana_port"`
-
-	// PrometheusPort is the host port Prometheus binds to (default: 9090).
-	PrometheusPort int `mapstructure:"prometheus_port"`
-
-	// LokiPort is the host port Loki binds to (default: 3100).
-	LokiPort int `mapstructure:"loki_port"`
-
-	// RetentionDays is how long Loki retains log data (default: 7).
-	RetentionDays int `mapstructure:"retention_days"`
-}
-
-// AuditConfig holds settings for the security audit log sink.
-type AuditConfig struct {
-	// Enabled toggles the audit log sink (default: true).
-	// When true, every security-relevant event is written to the configured output.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Output selects the write destination.
-	// Accepted values:
-	//   "stdout"       — write JSONL to standard output (default)
-	//   <file path>   — append JSONL to the file at the given path; the file is
-	//                   created if it does not exist.
-	// When Output is empty, "stdout" is assumed.
-	Output string `mapstructure:"output"`
-}
-
-// ErrorPagesConfig holds configuration for custom error page responses.
-// When enabled, VibeWarden serves files from Directory instead of the default
-// JSON error body for matching HTTP status codes.
-//
-// File naming convention: <status_code>.<ext> (e.g., 401.html, 403.json, 429.html).
-// Content-Type is inferred from the file extension:
-//   - .html  → text/html
-//   - .json  → application/json
-//
-// When no file matches a given status code, VibeWarden falls back to the
-// built-in JSON error response.
-type ErrorPagesConfig struct {
-	// Enabled toggles the custom error pages feature (default: false).
-	Enabled bool `mapstructure:"enabled"`
-
-	// Directory is the path to the directory containing custom error page files.
-	// The directory must be readable at startup. Required when Enabled is true.
-	Directory string `mapstructure:"directory"`
-}
-
-// MaintenanceConfig holds all settings for the maintenance mode plugin.
-// It maps to the maintenance section of vibewarden.yaml.
-type MaintenanceConfig struct {
-	// Enabled toggles maintenance mode (default: false).
-	// When true, all requests except those to /_vibewarden/* paths receive
-	// a 503 Service Unavailable response.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Message is the human-readable message returned to clients in the 503 body.
-	// Defaults to "Service is under maintenance" when empty.
-	Message string `mapstructure:"message"`
-}
-
-// CompressionConfig holds settings for response body compression.
-// Maps to the compression section of vibewarden.yaml.
-type CompressionConfig struct {
-	// Enabled toggles response compression. Default: true.
-	Enabled bool `mapstructure:"enabled"`
-
-	// Algorithms is the ordered list of compression algorithms to offer.
-	// Caddy negotiates the best match with the client via Accept-Encoding.
-	// Valid values: "gzip", "zstd".
-	// Default: ["zstd", "gzip"].
-	Algorithms []string `mapstructure:"algorithms"`
-}
-
-// ResponseHeadersConfig holds settings for arbitrary response header modification.
-// Maps to the response_headers section of vibewarden.yaml.
-// Operations are applied in the order: remove, then set, then add.
-type ResponseHeadersConfig struct {
-	// Set maps header names to values that overwrite any existing value (or create
-	// the header when absent). Values may reference environment variables using
-	// ${VAR} syntax; Caddy resolves these at request time.
-	Set map[string]string `mapstructure:"set"`
-
-	// Add maps header names to values that are appended to any existing value (or
-	// create the header when absent). Values may reference environment variables
-	// using ${VAR} syntax.
-	Add map[string]string `mapstructure:"add"`
-
-	// Remove is the list of header names to delete from every response.
-	Remove []string `mapstructure:"remove"`
-}
-
 // Validate checks the loaded configuration for logical consistency.
 // It returns a combined error listing all violations found.
 // Call Validate after Load to catch misconfiguration early.
@@ -1725,605 +265,22 @@ func (c *Config) Validate() error {
 		errs = append(errs, fmt.Sprintf("profile must be 'dev', 'tls', or 'prod', got %q", c.Profile))
 	}
 
-	// tls.provider validation: must be one of the accepted values.
-	// "acme" is accepted as an alias for "letsencrypt"; Load() normalises it
-	// before Validate() runs, but direct callers of Validate() may still pass
-	// it, so we permit it here as well.
-	switch c.TLS.Provider {
-	case "", "self-signed", "letsencrypt", "acme", "external",
-		"zerossl", "buypass", "letsencrypt-staging":
-		// valid — empty string is accepted (defaults to "self-signed" via Load)
-	default:
-		errs = append(errs, fmt.Sprintf(
-			"tls.provider %q is invalid; accepted values: \"self-signed\", \"letsencrypt\" (or alias \"acme\"), "+
-				"\"zerossl\", \"buypass\", \"letsencrypt-staging\", \"external\" — "+
-				"set tls.provider to one of those values",
-			c.TLS.Provider,
-		))
-	}
-
-	// ACME providers require a domain for certificate issuance.
-	// Also checked for "acme" — the alias — in case Validate() is called
-	// before Load() has had a chance to normalise the value.
-	acmeProviders := map[string]bool{
-		"letsencrypt": true, "acme": true,
-		"zerossl": true, "buypass": true, "letsencrypt-staging": true,
-	}
-	if c.TLS.Enabled && acmeProviders[c.TLS.Provider] && c.TLS.Domain == "" {
-		errs = append(errs, fmt.Sprintf(
-			"tls.domain is required when tls.provider is %q — "+
-				"set tls.domain to a domain you control and have pointed at this server "+
-				"(Let's Encrypt rejects reserved names like example.com)",
-			c.TLS.Provider,
-		))
-	}
-
-	// ZeroSSL requires email for automatic EAB registration.
-	if c.TLS.Enabled && c.TLS.Provider == "zerossl" && c.TLS.Email == "" {
-		errs = append(errs, "tls.email is required when tls.provider is \"zerossl\" — "+
-			"ZeroSSL needs an email for automatic EAB registration")
-	}
-
-	// TLS external provider requires cert_path and key_path.
-	if c.TLS.Enabled && c.TLS.Provider == "external" {
-		if c.TLS.CertPath == "" {
-			errs = append(errs, "tls.cert_path is required when tls.provider is \"external\" — "+
-				"set tls.cert_path to the path of your PEM-encoded certificate file")
-		}
-		if c.TLS.KeyPath == "" {
-			errs = append(errs, "tls.key_path is required when tls.provider is \"external\" — "+
-				"set tls.key_path to the path of your PEM-encoded private key file")
-		}
-	}
-
-	// Social providers: validate each entry.
-	for i, sp := range c.Auth.SocialProviders {
-		prefix := fmt.Sprintf("social_providers[%d]", i)
-
-		if !SupportedSocialProviders[sp.Provider] {
-			errs = append(errs, fmt.Sprintf("%s.provider %q is not supported; accepted values: google, github, apple, facebook, microsoft, gitlab, discord, slack, spotify, oidc", prefix, sp.Provider))
-		}
-		if sp.ClientID == "" {
-			errs = append(errs, fmt.Sprintf("%s.client_id is required", prefix))
-		}
-		if sp.ClientSecret == "" {
-			errs = append(errs, fmt.Sprintf("%s.client_secret is required", prefix))
-		}
-		if sp.Provider == "apple" {
-			if sp.TeamID == "" {
-				errs = append(errs, fmt.Sprintf("%s.team_id is required for provider \"apple\"", prefix))
-			}
-			if sp.KeyID == "" {
-				errs = append(errs, fmt.Sprintf("%s.key_id is required for provider \"apple\"", prefix))
-			}
-		}
-		if sp.Provider == "oidc" {
-			if sp.ID == "" {
-				errs = append(errs, fmt.Sprintf("%s.id is required for provider \"oidc\"", prefix))
-			}
-			if sp.IssuerURL == "" {
-				errs = append(errs, fmt.Sprintf("%s.issuer_url is required for provider \"oidc\"", prefix))
-			}
-		}
-	}
-
-	// auth.mode validation. Empty string is accepted here for robustness
-	// (SetDefault ensures "none" is used in practice).
-	switch c.Auth.Mode {
-	case "", AuthModeKratos, AuthModeJWT, AuthModeAPIKey, AuthModeNone:
-		// valid
-	default:
-		errs = append(errs, fmt.Sprintf(
-			"auth.mode %q is invalid; accepted values: \"none\", \"kratos\", \"jwt\", \"api-key\" — "+
-				"set auth.mode to one of those values (use \"none\" to disable authentication)",
-			c.Auth.Mode,
-		))
-	}
-
-	// auth.jwt validation (only when mode is "jwt").
-	if c.Auth.Mode == AuthModeJWT {
-		jwt := c.Auth.JWT
-		devJWKSMode := jwt.JWKSURL == "" && jwt.IssuerURL == ""
-
-		// In prod profile, local dev JWKS mode is not allowed — a real JWKS URL
-		// or OIDC issuer URL is required for production deployments.
-		if devJWKSMode && c.Profile == "prod" {
-			errs = append(errs, "auth.jwt: either jwks_url or issuer_url is required when auth.mode is \"jwt\" and profile is \"prod\"")
-		}
-
-		// Issuer and audience are required when a real JWKS source is configured.
-		// In dev JWKS mode they default to "vibewarden-dev" and "dev" respectively.
-		if !devJWKSMode {
-			if jwt.Issuer == "" {
-				errs = append(errs, "auth.jwt.issuer is required when auth.mode is \"jwt\"")
-			}
-			if jwt.Audience == "" {
-				errs = append(errs, "auth.jwt.audience is required when auth.mode is \"jwt\"")
-			}
-		}
-	}
-
-	// kratos.external validation: when external is true, public_url and admin_url
-	// must point to the external instance (they cannot be empty).
-	if c.Auth.Mode == AuthModeKratos && c.Kratos.External {
-		if c.Kratos.PublicURL == "" {
-			errs = append(errs, "kratos.public_url is required when kratos.external is true")
-		}
-		if c.Kratos.AdminURL == "" {
-			errs = append(errs, "kratos.admin_url is required when kratos.external is true")
-		}
-	}
-
-	// auth.on_kratos_unavailable validation.
-	if c.Auth.OnKratosUnavailable != "" &&
-		c.Auth.OnKratosUnavailable != "503" &&
-		c.Auth.OnKratosUnavailable != "allow_public" {
-		errs = append(errs, fmt.Sprintf(
-			"auth.on_kratos_unavailable %q is invalid; accepted values: \"503\", \"allow_public\"",
-			c.Auth.OnKratosUnavailable,
-		))
-	}
-
-	// auth.role_paths validation: only valid when mode is "kratos".
-	if len(c.Auth.RolePaths) > 0 && c.Auth.Mode != AuthModeKratos {
-		errs = append(errs, "auth.role_paths is only valid when auth.mode is \"kratos\"")
-	}
-	// Validate that role names in role_paths are recognised values.
-	// Delegates to identity.NewRole to keep the valid-role list in one place.
-	for roleName, paths := range c.Auth.RolePaths {
-		if _, err := identity.NewRole(roleName); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"auth.role_paths: role %q is invalid; accepted values: user, admin, moderator",
-				roleName,
-			))
-		}
-		if len(paths) == 0 {
-			errs = append(errs, fmt.Sprintf("auth.role_paths.%s: must have at least one path", roleName))
-		}
-	}
-
-	// Auth UI validation.
-	ui := c.Auth.UI
-	if ui.Mode != "" && ui.Mode != "built-in" && ui.Mode != "custom" {
-		errs = append(errs, fmt.Sprintf("auth.ui.mode %q is invalid; accepted values: \"built-in\", \"custom\"", ui.Mode))
-	}
-	if ui.PrimaryColor != "" && !hexColorRE.MatchString(ui.PrimaryColor) {
-		errs = append(errs, fmt.Sprintf("auth.ui.primary_color %q is not a valid hex color (expected #RGB or #RRGGBB)", ui.PrimaryColor))
-	}
-	if ui.BackgroundColor != "" && !hexColorRE.MatchString(ui.BackgroundColor) {
-		errs = append(errs, fmt.Sprintf("auth.ui.background_color %q is not a valid hex color (expected #RGB or #RRGGBB)", ui.BackgroundColor))
-	}
-	if ui.Mode == "custom" && ui.LoginURL == "" {
-		errs = append(errs, "auth.ui.login_url is required when auth.ui.mode is \"custom\"")
-	}
-
-	// body_size.max validation.
-	if c.BodySize.Max != "" {
-		if _, err := ParseBodySize(c.BodySize.Max); err != nil {
-			errs = append(errs, fmt.Sprintf("body_size.max: %s", err.Error()))
-		}
-	}
-
-	// ip_filter.mode validation.
-	if c.IPFilter.Enabled {
-		switch c.IPFilter.Mode {
-		case "", "allowlist", "blocklist":
-			// valid
-		default:
-			errs = append(errs, fmt.Sprintf(
-				"ip_filter.mode %q is invalid; accepted values: \"allowlist\", \"blocklist\"",
-				c.IPFilter.Mode,
-			))
-		}
-	}
-
-	// body_size.overrides validation.
-	for i, ov := range c.BodySize.Overrides {
-		prefix := fmt.Sprintf("body_size.overrides[%d]", i)
-		if ov.Path == "" {
-			errs = append(errs, fmt.Sprintf("%s.path is required", prefix))
-		}
-		if ov.Max != "" {
-			if _, err := ParseBodySize(ov.Max); err != nil {
-				errs = append(errs, fmt.Sprintf("%s.max: %s", prefix, err.Error()))
-			}
-		}
-	}
-
-	// webhooks.endpoints validation.
-	validFormats := map[string]bool{"": true, "raw": true, "slack": true, "discord": true}
-	for i, ep := range c.Webhooks.Endpoints {
-		prefix := fmt.Sprintf("webhooks.endpoints[%d]", i)
-		if ep.URL == "" {
-			errs = append(errs, fmt.Sprintf("%s.url is required", prefix))
-		}
-		if len(ep.Events) == 0 {
-			errs = append(errs, fmt.Sprintf("%s.events must have at least one entry", prefix))
-		}
-		if !validFormats[ep.Format] {
-			errs = append(errs, fmt.Sprintf("%s.format %q is invalid; accepted values: \"raw\", \"slack\", \"discord\"", prefix, ep.Format))
-		}
-		if ep.TimeoutSeconds < 0 {
-			errs = append(errs, fmt.Sprintf("%s.timeout_seconds must be >= 0", prefix))
-		}
-	}
-
-	// webhooks.signature_verification.paths validation.
-	validProviders := map[string]bool{
-		"stripe": true, "github": true, "slack": true, "twilio": true, "generic": true,
-	}
-	if c.Webhooks.SignatureVerification.Enabled {
-		for i, p := range c.Webhooks.SignatureVerification.Paths {
-			prefix := fmt.Sprintf("webhooks.signature_verification.paths[%d]", i)
-			if p.Path == "" {
-				errs = append(errs, fmt.Sprintf("%s.path is required", prefix))
-			}
-			if !validProviders[p.Provider] {
-				errs = append(errs, fmt.Sprintf(
-					"%s.provider %q is invalid; accepted values: stripe, github, slack, twilio, generic",
-					prefix, p.Provider,
-				))
-			}
-			if p.SecretEnvVar == "" {
-				errs = append(errs, fmt.Sprintf("%s.secret_env_var is required", prefix))
-			}
-			if p.Provider == "generic" && p.Header == "" {
-				errs = append(errs, fmt.Sprintf(
-					"%s.header is required when provider is \"generic\"", prefix,
-				))
-			}
-		}
-	}
-
-	// rate_limit.store validation.
-	switch c.RateLimit.Store {
-	case "", "memory":
-		// valid — "memory" is the default
-	case "redis":
-		if c.RateLimit.Redis.URL == "" && c.RateLimit.Redis.Address == "" {
-			errs = append(errs, "rate_limit.redis.address is required when rate_limit.store is \"redis\" and rate_limit.redis.url is not set — "+
-				"set rate_limit.redis.address to your Redis host:port (e.g. \"127.0.0.1:6379\") or set rate_limit.redis.url to a redis:// URL")
-		}
-		if c.RateLimit.Redis.URL != "" {
-			if err := validateRedisURL(c.RateLimit.Redis.URL); err != nil {
-				errs = append(errs, fmt.Sprintf("rate_limit.redis.url: %s", err.Error()))
-			}
-		}
-	default:
-		errs = append(errs, fmt.Sprintf(
-			"rate_limit.store %q is invalid; accepted values: \"memory\", \"redis\" — "+
-				"set rate_limit.store to \"memory\" for a single-process limiter or \"redis\" for a distributed limiter",
-			c.RateLimit.Store,
-		))
-	}
-
-	// telemetry.logs.otlp requires telemetry.otlp.endpoint.
-	if c.Telemetry.Logs.OTLP && c.Telemetry.OTLP.Endpoint == "" {
-		errs = append(errs, "telemetry.logs.otlp requires telemetry.otlp.endpoint")
-	}
-
-	// upstream.health validation.
-	if c.Upstream.Health.Enabled {
-		if c.Upstream.Health.Path == "" {
-			errs = append(errs, "upstream.health.path must not be empty when upstream.health.enabled is true")
-		}
-		if c.Upstream.Health.Interval != "" {
-			if _, err := time.ParseDuration(c.Upstream.Health.Interval); err != nil {
-				errs = append(errs, fmt.Sprintf("upstream.health.interval %q is not a valid duration: %s", c.Upstream.Health.Interval, err.Error()))
-			}
-		}
-		if c.Upstream.Health.Timeout != "" {
-			if _, err := time.ParseDuration(c.Upstream.Health.Timeout); err != nil {
-				errs = append(errs, fmt.Sprintf("upstream.health.timeout %q is not a valid duration: %s", c.Upstream.Health.Timeout, err.Error()))
-			}
-		}
-		if c.Upstream.Health.UnhealthyThreshold <= 0 {
-			errs = append(errs, fmt.Sprintf(
-				"upstream.health.unhealthy_threshold %d is invalid; must be > 0",
-				c.Upstream.Health.UnhealthyThreshold,
-			))
-		}
-		if c.Upstream.Health.HealthyThreshold <= 0 {
-			errs = append(errs, fmt.Sprintf(
-				"upstream.health.healthy_threshold %d is invalid; must be > 0",
-				c.Upstream.Health.HealthyThreshold,
-			))
-		}
-	}
-
-	// cors validation.
-	if c.CORS.Enabled && c.CORS.AllowCredentials {
-		for _, o := range c.CORS.AllowedOrigins {
-			if o == "*" {
-				errs = append(errs, "cors.allow_credentials: true cannot be combined with cors.allowed_origins: [\"*\"]; browsers reject credentialed requests to wildcard origins")
-				break
-			}
-		}
-	}
-
-	// observability validation.
-	if c.Observability.Enabled {
-		if c.Observability.GrafanaPort <= 0 || c.Observability.GrafanaPort > 65535 {
-			errs = append(errs, fmt.Sprintf(
-				"observability.grafana_port %d is invalid; must be 1-65535",
-				c.Observability.GrafanaPort,
-			))
-		}
-		if c.Observability.PrometheusPort <= 0 || c.Observability.PrometheusPort > 65535 {
-			errs = append(errs, fmt.Sprintf(
-				"observability.prometheus_port %d is invalid; must be 1-65535",
-				c.Observability.PrometheusPort,
-			))
-		}
-		if c.Observability.LokiPort <= 0 || c.Observability.LokiPort > 65535 {
-			errs = append(errs, fmt.Sprintf(
-				"observability.loki_port %d is invalid; must be 1-65535",
-				c.Observability.LokiPort,
-			))
-		}
-		if c.Observability.RetentionDays <= 0 {
-			errs = append(errs, fmt.Sprintf(
-				"observability.retention_days %d is invalid; must be > 0",
-				c.Observability.RetentionDays,
-			))
-		}
-	}
-
-	// database.external_url validation: when set, must be a valid postgres:// URL.
-	if c.Database.ExternalURL != "" {
-		if err := validatePostgresURL(c.Database.ExternalURL); err != nil {
-			errs = append(errs, fmt.Sprintf("database.external_url: %s", err.Error()))
-		}
-	}
-
-	// database.tls_mode validation.
-	validTLSModes := map[string]bool{
-		"":            true, // empty means use default ("require")
-		"disable":     true,
-		"require":     true,
-		"verify-ca":   true,
-		"verify-full": true,
-	}
-	if !validTLSModes[c.Database.TLSMode] {
-		errs = append(errs, fmt.Sprintf(
-			"database.tls_mode %q is invalid; accepted values: \"disable\", \"require\", \"verify-ca\", \"verify-full\"",
-			c.Database.TLSMode,
-		))
-	}
-
-	// database.pool validation.
-	if c.Database.Pool.MaxConns < 0 {
-		errs = append(errs, fmt.Sprintf(
-			"database.pool.max_conns %d is invalid; must be >= 0",
-			c.Database.Pool.MaxConns,
-		))
-	}
-	if c.Database.Pool.MinConns < 0 {
-		errs = append(errs, fmt.Sprintf(
-			"database.pool.min_conns %d is invalid; must be >= 0",
-			c.Database.Pool.MinConns,
-		))
-	}
-	if c.Database.Pool.MaxConns > 0 && c.Database.Pool.MinConns > c.Database.Pool.MaxConns {
-		errs = append(errs, fmt.Sprintf(
-			"database.pool.min_conns (%d) must be <= database.pool.max_conns (%d)",
-			c.Database.Pool.MinConns, c.Database.Pool.MaxConns,
-		))
-	}
-
-	// database.connect_timeout validation.
-	if c.Database.ConnectTimeout != "" {
-		if _, err := time.ParseDuration(c.Database.ConnectTimeout); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"database.connect_timeout %q is not a valid duration: %s",
-				c.Database.ConnectTimeout, err.Error(),
-			))
-		}
-	}
-
-	// egress validation.
+	errs = append(errs, validateTLS(c)...)
+	errs = append(errs, validateAuth(c)...)
+	errs = append(errs, validateSidecar(c)...)
+	errs = append(errs, validateWebhooks(c)...)
+	errs = append(errs, validateRateLimit(c)...)
+	errs = append(errs, validateTelemetry(c)...)
+	errs = append(errs, validateUpstream(c)...)
+	errs = append(errs, validateCORS(c)...)
+	errs = append(errs, validateObservability(c)...)
+	errs = append(errs, validateDatabase(c)...)
 	if c.Egress.Enabled {
-		egressErrs := validateEgressConfig(c.Egress)
-		errs = append(errs, egressErrs...)
-	}
-
-	// error_pages validation.
-	if c.ErrorPages.Enabled && c.ErrorPages.Directory == "" {
-		errs = append(errs, "error_pages.directory is required when error_pages.enabled is true")
+		errs = append(errs, validateEgressConfig(c.Egress)...)
 	}
 
 	if len(errs) > 0 {
 		return fmt.Errorf("%s", strings.Join(errs, "; "))
-	}
-	return nil
-}
-
-// validateEgressConfig validates the egress proxy configuration and returns
-// a slice of error strings (one per violation). An empty slice means the
-// configuration is valid.
-func validateEgressConfig(cfg EgressConfig) []string {
-	var errs []string
-
-	// dns.allowed_private CIDR validation.
-	for i, cidr := range cfg.DNS.AllowedPrivate {
-		if _, _, err := net.ParseCIDR(cidr); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"egress.dns.allowed_private[%d] %q is not a valid CIDR: %s",
-				i, cidr, err.Error(),
-			))
-		}
-	}
-
-	// default_policy validation.
-	switch cfg.DefaultPolicy {
-	case "", "deny", "allow":
-		// valid — empty falls back to "deny" via Load defaults
-	default:
-		errs = append(errs, fmt.Sprintf(
-			"egress.default_policy %q is invalid; accepted values: \"allow\", \"deny\"",
-			cfg.DefaultPolicy,
-		))
-	}
-
-	// default_timeout validation.
-	if cfg.DefaultTimeout != "" {
-		if _, err := time.ParseDuration(cfg.DefaultTimeout); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"egress.default_timeout %q is not a valid duration: %s",
-				cfg.DefaultTimeout, err.Error(),
-			))
-		}
-	}
-
-	// default_body_size_limit validation.
-	if cfg.DefaultBodySizeLimit != "" {
-		if _, err := ParseBodySize(cfg.DefaultBodySizeLimit); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"egress.default_body_size_limit: %s", err.Error(),
-			))
-		}
-	}
-
-	// default_response_size_limit validation.
-	if cfg.DefaultResponseSizeLimit != "" {
-		if _, err := ParseBodySize(cfg.DefaultResponseSizeLimit); err != nil {
-			errs = append(errs, fmt.Sprintf(
-				"egress.default_response_size_limit: %s", err.Error(),
-			))
-		}
-	}
-
-	// routes validation.
-	routeNames := make(map[string]bool, len(cfg.Routes))
-	for i, r := range cfg.Routes {
-		prefix := fmt.Sprintf("egress.routes[%d]", i)
-
-		if r.Name == "" {
-			errs = append(errs, fmt.Sprintf("%s.name is required", prefix))
-		} else if routeNames[r.Name] {
-			errs = append(errs, fmt.Sprintf("%s.name %q is a duplicate; route names must be unique", prefix, r.Name))
-		} else {
-			routeNames[r.Name] = true
-		}
-
-		if r.Pattern == "" {
-			errs = append(errs, fmt.Sprintf("%s.pattern is required", prefix))
-		} else {
-			if !strings.HasPrefix(r.Pattern, "http://") && !strings.HasPrefix(r.Pattern, "https://") {
-				errs = append(errs, fmt.Sprintf("%s.pattern %q must start with http:// or https://", prefix, r.Pattern))
-			}
-		}
-
-		if r.Timeout != "" {
-			if _, err := time.ParseDuration(r.Timeout); err != nil {
-				errs = append(errs, fmt.Sprintf(
-					"%s.timeout %q is not a valid duration: %s",
-					prefix, r.Timeout, err.Error(),
-				))
-			}
-		}
-
-		// circuit_breaker validation: both threshold and reset_after must be set together.
-		cb := r.CircuitBreaker
-		hasCBThreshold := cb.Threshold != 0
-		hasCBReset := cb.ResetAfter != ""
-		if hasCBThreshold || hasCBReset {
-			if cb.Threshold <= 0 {
-				errs = append(errs, fmt.Sprintf("%s.circuit_breaker.threshold must be > 0", prefix))
-			}
-			if cb.ResetAfter == "" {
-				errs = append(errs, fmt.Sprintf("%s.circuit_breaker.reset_after is required when circuit_breaker.threshold is set", prefix))
-			} else if _, err := time.ParseDuration(cb.ResetAfter); err != nil {
-				errs = append(errs, fmt.Sprintf(
-					"%s.circuit_breaker.reset_after %q is not a valid duration: %s",
-					prefix, cb.ResetAfter, err.Error(),
-				))
-			}
-		}
-
-		// retries validation.
-		if r.Retries.Max < 0 {
-			errs = append(errs, fmt.Sprintf("%s.retries.max must be >= 0", prefix))
-		}
-		switch r.Retries.Backoff {
-		case "", "exponential", "fixed":
-			// valid
-		default:
-			errs = append(errs, fmt.Sprintf(
-				"%s.retries.backoff %q is invalid; accepted values: \"exponential\", \"fixed\"",
-				prefix, r.Retries.Backoff,
-			))
-		}
-
-		// body_size_limit validation.
-		if r.BodySizeLimit != "" {
-			if _, err := ParseBodySize(r.BodySizeLimit); err != nil {
-				errs = append(errs, fmt.Sprintf("%s.body_size_limit: %s", prefix, err.Error()))
-			}
-		}
-
-		// response_size_limit validation.
-		if r.ResponseSizeLimit != "" {
-			if _, err := ParseBodySize(r.ResponseSizeLimit); err != nil {
-				errs = append(errs, fmt.Sprintf("%s.response_size_limit: %s", prefix, err.Error()))
-			}
-		}
-
-		// secret injection: if any secret field is set, name and header are required.
-		if r.SecretFormat != "" || r.SecretHeader != "" {
-			if r.Secret == "" {
-				errs = append(errs, fmt.Sprintf("%s.secret is required when secret_header or secret_format is set", prefix))
-			}
-			if r.SecretHeader == "" {
-				errs = append(errs, fmt.Sprintf("%s.secret_header is required when secret or secret_format is set", prefix))
-			}
-		}
-
-		// prompt_injection validation.
-		if r.PromptInjection.Enabled {
-			switch r.PromptInjection.Action {
-			case "", "block", "detect":
-				// valid — empty defaults to "block"
-			default:
-				errs = append(errs, fmt.Sprintf(
-					"%s.prompt_injection.action %q is invalid; accepted values: \"block\", \"detect\"",
-					prefix, r.PromptInjection.Action,
-				))
-			}
-		}
-	}
-
-	return errs
-}
-
-// validatePostgresURL returns an error if s is not a valid postgres:// URL.
-// It accepts both "postgres://" and "postgresql://" schemes.
-func validatePostgresURL(s string) error {
-	u, err := url.Parse(s)
-	if err != nil {
-		return fmt.Errorf("must be a valid URL: %w", err)
-	}
-	if u.Scheme != "postgres" && u.Scheme != "postgresql" {
-		return fmt.Errorf("must use postgres:// or postgresql:// scheme, got %q", u.Scheme)
-	}
-	if u.Host == "" {
-		return fmt.Errorf("must include a host")
-	}
-	return nil
-}
-
-// validateRedisURL returns an error if s is not a valid Redis connection URL.
-// It accepts "redis://" (plain) and "rediss://" (TLS) schemes.
-func validateRedisURL(s string) error {
-	u, err := url.Parse(s)
-	if err != nil {
-		return fmt.Errorf("must be a valid URL: %w", err)
-	}
-	if u.Scheme != "redis" && u.Scheme != "rediss" {
-		return fmt.Errorf("must use redis:// or rediss:// scheme, got %q", u.Scheme)
-	}
-	if u.Host == "" {
-		return fmt.Errorf("must include a host")
 	}
 	return nil
 }
